+++ frontend/workflows/image_gen/config_panel.js
@@ -1,14 +1,10 @@
+import { api, closeModal, esc, escAttr, registerAction, showModal, toast } from "/static/workflow_api.js";
 import {
-  api,
-  closeModal,
-  convUrl,
-  esc,
-  escAttr,
-  getActiveConvId,
-  registerAction,
-  showModal,
-  toast,
-} from "/static/workflow_api.js";
+  initCharacterProfile,
+  populateProfile,
+  resetCharacterProfile,
+  saveProfile,
+} from "./character_profile.js";
 import {
   classTypes,
   graphFromApiJson,
@@ -47,8 +43,6 @@ const REFERENCE_SOURCES = [
 // image that previews fine and is silently gone on the next open.
 const MAX_REFERENCE_SLOTS = 4;
 const MAX_USER_GRAPHS = 32;
-const MAX_REFERENCE_IMAGE_BYTES = 10_000_000;
-const REFERENCE_IMAGE_MIMES = ["image/png", "image/jpeg", "image/webp"];
 
 // Resolution presets for the cloud picker, stored as pixels even for providers that
 // speak aspect ratios — one canonical representation, converted at the wire by a
@@ -124,10 +118,6 @@ export function initConfigPanel(sharedConfig) {
   registerAction(WORKFLOW_ID, "settingsClose", () => closeModal());
   registerAction(WORKFLOW_ID, "save", () => saveSettings());
   registerAction(WORKFLOW_ID, "graphFile", (el) => importGraphFile(el));
-  registerAction(WORKFLOW_ID, "referenceFile", (el) => pickReferenceImage(el));
-  registerAction(WORKFLOW_ID, "referenceClear", () =>
-    setReferenceImage({ reference_image_b64: "", reference_mime: "" }),
-  );
   registerAction(WORKFLOW_ID, "graphAdd", () => addPendingGraph());
   registerAction(WORKFLOW_ID, "graphRemove", (el) => removeGraph(el.dataset.graphId));
   registerAction(WORKFLOW_ID, "styleAdd", () => addStyle());
@@ -139,6 +129,7 @@ export function initConfigPanel(sharedConfig) {
   registerAction(WORKFLOW_ID, "connChange", (el) => refreshConnectionState(el));
   registerAction(WORKFLOW_ID, "connTest", (el) => testConnection(el.dataset.connId));
   registerAction(WORKFLOW_ID, "connOpen", (el) => revealConnection(el.dataset.connId));
+  initCharacterProfile();
 }
 
 // Every conversation-less config/discovery call rides the one QUERY route. Not
@@ -537,9 +528,8 @@ function openStyleIds() {
     .filter(Boolean);
 }
 
-// The permanent gaps, stated once under the connection rather than on every render:
-// a note that fires 100% of the time is one users learn to skip, which then hides
-// the per-render disclosures that actually vary.
+// The permanent gaps, stated once under the connection rather than on every render —
+// see the `gaps` field on ProviderPreset for why they are not per-render notes.
 function capabilityLine(preset) {
   if (!preset) return "";
   const gaps = Array.isArray(preset.gaps) ? preset.gaps : [];
@@ -725,11 +715,19 @@ function captureConnections() {
   }
 }
 
+// Both halves of the form, together. Every caller that re-renders needs both — a
+// style row and a connection row can each hold an uncommitted edit — so pairing
+// them here is what stops a new call site remembering only one and dropping the
+// other's live values on the next innerHTML swap.
+function captureForm() {
+  captureStyles();
+  captureConnections();
+}
+
 function addConnection() {
   const id = document.getElementById("ig-conn-add")?.value;
   if (!id) return;
-  captureStyles();
-  captureConnections();
+  captureForm();
   const preset = providerFor(id);
   const existing = draft.connections[id] || {};
   draft.connections[id] = {
@@ -753,8 +751,7 @@ function addConnection() {
 }
 
 function removeConnection(id) {
-  captureStyles();
-  captureConnections();
+  captureForm();
   delete draft.connections[id];
   delete modelsByConnection[id];
   pendingConnections.delete(id);
@@ -776,8 +773,7 @@ function refreshConnectionState(el) {
   const row = el.closest("details.ig-conn");
   const id = row?.dataset.connId;
   if (!id) return;
-  captureStyles();
-  captureConnections();
+  captureForm();
   rebuildConnections();
   const connection = findConnection(connections, id);
   const detail = row.querySelector(".ig-conn-detail");
@@ -820,8 +816,7 @@ function configForConnection(id) {
 // survive the asynchronous re-render.
 function applyModels(id, names) {
   const openStyles = openStyleIds();
-  captureStyles();
-  captureConnections();
+  captureForm();
   modelsByConnection[id] = modelPickerState(names).models;
   rebuildConnections();
   if (id === COMFY_CONNECTION) {
@@ -887,7 +882,7 @@ function openSettings(expandStyleId = "") {
   probeIds = {};
   modelsByConnection = {};
   pendingConnections = new Set();
-  referenceImage = { reference_image_b64: "", reference_mime: "" };
+  resetCharacterProfile();
   draft = {
     styles: (Array.isArray(cfg.styles) ? cfg.styles : []).map((s) => ({ ...s })),
     graphs: (Array.isArray(ext.user_graphs) ? ext.user_graphs : []).map((g) => ({ ...g })),
@@ -945,8 +940,7 @@ function openSettings(expandStyleId = "") {
 }
 
 function readConfig() {
-  captureStyles();
-  captureConnections();
+  captureForm();
   const ext = cfg.external_comfy || {};
   return {
     // `source`, `cloud.provider`, `default_style` and `pov_mode` have no control in
@@ -1145,110 +1139,3 @@ function confirmRemotePrivacy(next) {
   return true;
 }
 
-// The character's reference image as the form holds it — loaded with the profile,
-// replaced by the picker, emptied by Clear, written back on Save. Module state
-// rather than read off the rendered <img>, so a save that never touched the picker
-// round-trips the stored bytes untouched.
-let referenceImage = { reference_image_b64: "", reference_mime: "" };
-
-function referenceImageHtml() {
-  const stored = !!referenceImage.reference_image_b64;
-  return `<div class="ig-reference-image">
-      ${stored ? `<div class="ig-reference-preview"><img class="ig-reference-thumb" alt="Character reference image" src="data:${escAttr(referenceImage.reference_mime || "image/png")};base64,${escAttr(referenceImage.reference_image_b64)}"></div>` : ""}
-      <div class="ig-reference-controls">
-        <input type="file" accept="image/png,image/jpeg,image/webp" data-wf-action="image_gen:referenceFile" data-wf-on="change">
-        ${
-          stored
-            ? `<button class="btn btn-sm" data-wf-action="image_gen:referenceClear">Clear</button>`
-            : `<span class="image-gen-note ig-reference-empty">No reference image — the character card's avatar is used.</span>`
-        }
-      </div>
-    </div>`;
-}
-
-function setReferenceImage(next) {
-  referenceImage = next;
-  const host = document.getElementById("ig-reference-host");
-  if (host) host.innerHTML = referenceImageHtml();
-}
-
-async function pickReferenceImage(input) {
-  const file = input.files?.[0];
-  if (!file) return;
-  if (file.size > MAX_REFERENCE_IMAGE_BYTES) {
-    toast("That image is too large — use one under 10 MB", "error");
-    input.value = "";
-    return;
-  }
-  if (!REFERENCE_IMAGE_MIMES.includes((file.type || "").toLowerCase())) {
-    toast("Orb accepts PNG, JPEG and WebP reference images", "error");
-    input.value = "";
-    return;
-  }
-  try {
-    // Chunked: a single spread of the whole array blows String.fromCharCode's
-    // argument limit on a multi-MB image.
-    const bytes = new Uint8Array(await file.arrayBuffer());
-    let binary = "";
-    for (let i = 0; i < bytes.length; i += 0x8000) binary += String.fromCharCode(...bytes.subarray(i, i + 0x8000));
-    setReferenceImage({ reference_image_b64: btoa(binary), reference_mime: file.type.toLowerCase() });
-  } catch {
-    toast("Could not read that image", "error");
-  }
-  input.value = "";
-}
-
-async function populateProfile() {
-  const el = document.getElementById("ig-profile");
-  if (!el || !getActiveConvId()) return;
-  try {
-    const res = await api.post(convUrl(getActiveConvId(), "workflows", WORKFLOW_ID, "trigger"), {
-      action: "get_profile",
-    });
-    if (!res?.profile) {
-      el.textContent = "This conversation has no character.";
-      return;
-    }
-    el.classList.remove("image-gen-note");
-    referenceImage = {
-      reference_image_b64: res.profile.reference_image_b64 || "",
-      reference_mime: res.profile.reference_mime || "",
-    };
-    el.innerHTML = `<div class="ig-profile-fields">
-        <label>Positive prompt<textarea id="ig-appearance" placeholder="Permanent tags, fill with permanent traits (e.g. Hatsune Miku, black and white)">${esc(res.profile.appearance_prompt || "")}</textarea></label>
-        <label>Negative prompt<textarea id="ig-profile-negative" placeholder="Things to never render (e.g. 3D, colored, color). Quality and scene negatives are already handled.">${esc(res.profile.negative_prompt || "")}</textarea></label>
-        <div class="ig-profile-reference">
-          <span class="ig-profile-reference-label">Reference image</span>
-          <span class="image-gen-note">Used by workflows with reference image slots.</span>
-          <div id="ig-reference-host">${referenceImageHtml()}</div>
-        </div>
-      </div>`;
-  } catch {
-    el.textContent = "Could not load character appearance.";
-  }
-}
-
-async function saveProfile() {
-  // No fields rendered means no active character: sending blanks would wipe a
-  // saved appearance.
-  const appearanceEl = document.getElementById("ig-appearance");
-  if (!appearanceEl || !getActiveConvId()) return;
-  const res = await api.post(convUrl(getActiveConvId(), "workflows", WORKFLOW_ID, "trigger"), {
-    action: "set_profile",
-    profile: {
-      appearance_prompt: appearanceEl.value || "",
-      negative_prompt: document.getElementById("ig-profile-negative")?.value || "",
-      ...referenceImage,
-    },
-  });
-  // A save that reports success while discarding what the form is still previewing
-  // is the one outcome the user cannot diagnose, so the handler's warning is shown
-  // and the local copy is brought back in line with what was stored.
-  if (res?.warning) {
-    toast(res.warning, "error");
-    referenceImage = {
-      reference_image_b64: res.profile?.reference_image_b64 || "",
-      reference_mime: res.profile?.reference_mime || "",
-    };
-  }
-}
